Corrigiendo casteo innecesario - ejercicio:2 clase:2

+++ src/main/java/org/example/Main.java
@@ -7,13 +7,16 @@ import static java.lang.Math.sqrt;
 public class Main {
     public static void main(String[] args) {
         Scanner keyboardInput = new Scanner(System.in);
-        // Segundo ejercicio propuesto - clase 2 ( e:2 c:2)
+        // Tercer ejercicio propuesto - clase 2 ( e:3 c:2 )
+
+
+        // Segundo ejercicio propuesto - clase 2 ( e:2 c:2 )
         int numberGiven;
         double squareRoot;
 
         System.out.print("Ingrese un número entero: ");
         numberGiven = keyboardInput.nextInt();
-        squareRoot = Math.sqrt((double)numberGiven);
+        squareRoot = Math.sqrt(numberGiven);
 
 
         System.out.println("La raíz cuadrada de "+ numberGiven+" es: "+ String.format("%.2f", squareRoot));
